requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile - -o requirements.txt
absl-py==2.1.0
    # via
    #   chex
    #   optax
antlr4-python3-runtime==4.9.3
    # via
    #   hydra-core
    #   omegaconf
beartype==0.18.5
certifi==2024.7.4
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.3.2
    # via requests
chex==0.1.86
    # via optax
click==8.1.7
    # via wandb
docker-pycreds==0.4.0
    # via wandb
equinox==0.11.5
etils==1.9.2
    # via optax
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via wandb
hydra-core==1.3.2
idna==3.7
    # via requests
iniconfig==2.0.0
    # via pytest
jax==0.4.31
    # via
    #   chex
    #   equinox
    #   optax
jax-cuda12-pjrt==0.4.31
    # via jax-cuda12-plugin
jax-cuda12-plugin==0.4.31
jaxlib==0.4.31
    # via
    #   chex
    #   jax
    #   optax
jaxtyping==0.2.33
    # via equinox
ml-dtypes==0.4.0
    # via
    #   jax
    #   jaxlib
networkx==3.3
numpy==2.1.0
    # via
    #   chex
    #   jax
    #   jaxlib
    #   ml-dtypes
    #   opt-einsum
    #   optax
    #   scipy
nvidia-cublas-cu12==12.6.0.22
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
nvidia-cuda-cupti-cu12==12.6.37
nvidia-cuda-nvcc-cu12==12.6.20
nvidia-cuda-runtime-cu12==12.6.37
nvidia-cudnn-cu12==9.3.0.75
nvidia-cufft-cu12==11.2.6.28
nvidia-cusolver-cu12==11.6.4.38
nvidia-cusparse-cu12==12.5.2.23
    # via nvidia-cusolver-cu12
nvidia-nccl-cu12==2.22.3
nvidia-nvjitlink-cu12==12.6.20
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
omegaconf==2.3.0
    # via hydra-core
opt-einsum==3.3.0
    # via jax
optax==0.2.3
packaging==24.1
    # via
    #   hydra-core
    #   pytest
pip==24.0
platformdirs==4.2.2
    # via wandb
pluggy==1.5.0
    # via pytest
protobuf==5.27.3
    # via wandb
psutil==6.0.0
    # via wandb
pytest==8.3.2
pyyaml==6.0.2
    # via
    #   omegaconf
    #   wandb
requests==2.32.3
    # via wandb
scipy==1.14.1
    # via
    #   jax
    #   jaxlib
sentry-sdk==2.13.0
    # via wandb
setproctitle==1.3.3
    # via wandb
setuptools==73.0.1
    # via
    #   chex
    #   wandb
six==1.16.0
    # via docker-pycreds
smmap==5.0.1
    # via gitdb
toolz==0.12.1
    # via chex
tqdm==4.66.5
typeguard==2.13.3
    # via jaxtyping
typing-extensions==4.12.2
    # via
    #   chex
    #   equinox
    #   etils
urllib3==2.2.2
    # via
    #   requests
    #   sentry-sdk
wandb==0.17.7
